fix issues with copier resolution
When a local path is provided, that is also a git repo it would match
the git copier, and proceed to do a git checkout ignoring the local path.

Additionally, when the path provided is an absolute path the "root_path"
which is used to allow relative links was being used by the copier match
logic resulting in an error.