INF-5381: implement a fallback method for getting remote state items
When using --limit, which is often standard practice, remote definitions are not setup.
This is generally acceptable, however when a hook script requires a remote state item
and the remote state is not setup, the previous mechanism would fail and not set the
expected TF_REMOTE_VAR environment variables.

In order to ensure that the TF_REMOTE_VAR's are always available, a fall back mechanism
of using `terraform show` to get remote state items was added, this is a slower method
the the original so the original is preferred.