merge upstream changes (#40)
* Refactor load_config in commands.root

* Update isort profile to be compatible with preferred formatter

* remove exception handling for FileNotFoundError

This exception is handled within the RootCommand object, so the
exception will never bubble up.

* refactor get_version out of a the Version command

The version should generally be set, not just when executing the version
command from the CLI. This commit refactors the get_version command into
system.util; and updates the corresponding tests.

This also adds additional test coverage for other minor commands.

* refactor the s3 backend

This refactoring included breaking down several of the long methods in
the s3 backend to smaller functions in order to enhance unit testing and
ensure that this section of code remains solid.

This also updates moto from v4.x to v5.x to take advantage of new
functionality and simplified use.

There are many test additions for the s3 backend, but also a lot of
changes to try and reduce the reliance on complicated fixtures and
provide smaller more consise tests.

* update deps to address security issues

This not only updates security specific deps, but does a broad update of
all deps. The only problem areas:

- awscli: This requires an old colorama, but awscli isn't actually a
dep, so remove it. It was only included to make running the AWSCLI
easier at a point earlier in the projects life

- pytest-lazyfixtures: This pytest plugin relies on deprecated
functionality and has not been updated in >2 years, so likely will not
be fixed soon. The only tests that relied on lazy fixutres were able to
be pretty quickly fixed by using the pytest request.getfixturevalue
method

---------

Co-authored-by: Richard Maynard <richard.maynard@gmail.com>